ENH Add integration test script and YAML axis
This PR works to move all of the config and scripts for all integration/nightly tests out of Jenkins and into the repo. In addition, using the `tests.yaml` axis config file will make updates to all tests easier as they all share a common YAML config file.

In this PR the scripts are not cleaned-up or consolidated; that will be done in a future PR.

In addition, this PR will remove CUDA 10.0 and Python 3.6 from the testing options in prep for dropping those combinations across RAPIDS for v0.15 release.