FIX Correct conda recipes to use RAPIDS version
Currently the recipes use `version` which is based off of the tag in this repo. Switching this to `rapids_version` which uses the env var `RAPIDS_VER` specified by `./ci/axis/nightly.yaml` or `./ci/axis/release.yaml`. This way during burndown where we need to build multiple version from the `nightly.yaml` axis they will each get their respective version.

As it is right now they get the same version so while a `0.16` build shows as successful it is versioned as `0.15` and has a duplicate package.