FIX Update RAPIDS versions in axis files
With #96 RAPIDS versions need to be full versions like `0.15.0a` instead of `0.15`. This is because we are no longer using the git tag to pull this information and without it we get conda files with the version `0.15` instead of matching the nightly version of `0.15.0a`.

Additionally cleaned up our use of git tags to set version as the axis files setting `RAPIDS_VER` will control the versioning in this repo.