FIX Add setuptools to rapids meta pkg
Adding `setuptools` pin to `rapids` meta-pkg to ensure the pin takes precedence over others. Previously we've relied on the pins with the dependencies of `rapids`, but there are conflicts with `cupy` and other packages that do not have a set pin. Having it as a direct dependency will make the solver respect the pin in all solves.