Update `dask` & `distributed` minimum versions
Currently when I run `conda create -n test-env rapids-build-env` locally, it installs an older version of `dask` and `distributed` which is causing issues in our `devel` image builds shown in the log link below. When I create a new environment with `conda create -n test-env rapids-build-env "dask>=2022.02.1"`, it correctly installs the latest `dask` and `distributed` versions and fixes the build issue. This PR updates the `dask` and `distributed` versions to match the working version specification.

- https://gpuci.gpuopenanalytics.com/job/rapidsai/job/docker/job/rapidsai-core-devel/1105/BUILD_IMAGE=rapidsai%2Frapidsai-core-dev-nightly,CUDA_VER=11.5,FROM_IMAGE=gpuci%2Frapidsai,IMAGE_TYPE=devel,LINUX_VER=ubuntu20.04,PYTHON_VER=3.8,RAPIDS_VER=22.04/console