Drop `nvidia` channel
Previously RAPIDS used the `nvidia` channel to get additional CUDA 11
packages that were only available from that channel. With the
restructuring of the Conda packages with CUDA 12 and the resyncing of
the `nvidia` and `conda-forge` channels, the same content can be
installed from either channel. To simplify things, stick with the
`conda-forge` channel as all other dependencies need to come from there
anyways.